+++ Model/Microsoft/Adult.cs
@@ -4,7 +4,7 @@
     {
         public bool IsAdultContent { get; set; }
         public bool IsRacyContent { get; set; }
-        public double AdultScore { get; set; }
-        public double RacyScore { get; set; }
+        public decimal AdultScore { get; set; }
+        public decimal RacyScore { get; set; }
     }
 }
